requirements.txt
torchvision
torchsummary
torch
tensorboardX
numpy
Pillow
Dominate
